requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas==2.0.3
python-dotenv
langchain-ollama
langchain-community
langchain-core
langchain-openai
langgraph